+++ Gw2Sharp/WebApi/V2/Clients/IBulkAliasExpandableClient.cs
@@ -1,0 +1,34 @@
+using System;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Gw2Sharp.WebApi.V2.Clients
+{
+    /// <summary>
+    /// Implements an endpoint client that is bulk expandable with an aliased type.
+    /// </summary>
+    /// <typeparam name="TObject">The response object type.</typeparam>
+    /// <typeparam name="TId">The id value type.</typeparam>
+    public interface IBulkAliasExpandableClient<TObject, TId> : IEndpointClient
+        where TObject : IApiV2Object
+    {
+        /// <summary>
+        /// Requests a single entry by id.
+        /// </summary>
+        /// <param name="id">The entry id.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The entry.</returns>
+        /// <exception cref="ArgumentNullException"><paramref name="id"/> is <c>null</c>.</exception>
+        Task<TObject> GetAsync(TId id, CancellationToken cancellationToken = default);
+
+        /// <summary>
+        /// Requests many entries by their id (a.k.a. bulk expansion).
+        /// </summary>
+        /// <param name="ids">The entry ids.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        /// <returns>The entries.</returns>
+        /// <exception cref="ArgumentNullException"><paramref name="ids"/> is <c>null</c>.</exception>
+        Task<IReadOnlyList<TObject>> ManyAsync(IEnumerable<TId> ids, CancellationToken cancellationToken = default);
+    }
+}